Implement a common dims simplifier.
